набросал структуру приложения
