+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,0 +1,12 @@
+package ru.yandex.practicum.filmorate.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class FilmService {
+    //добавление лайка
+    //удаление лайка
+    //вывод 10 наиболее популярных фильмов по количеству лайков
+    /*
+     * Пока пусть каждый пользователь может поставить лайк фильму только один раз.*/
+}
