+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,0 +1,12 @@
+package ru.yandex.practicum.filmorate.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+    //добавление в друзья
+    //удаление из друзей
+    //вывод списка общих друзей
+    /*
+     * Пока пользователям не надо одобрять заявки в друзья — добавляем сразу. То есть если Лена стала другом Саши, то это значит, что Саша теперь друг Лены.*/
+}
