+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -1,0 +1,15 @@
+package ru.yandex.practicum.filmorate.storage.film;
+
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.util.Collection;
+
+public interface FilmStorage {
+    Collection<Film> findAll();
+
+    Film create(Film film);
+
+    Film update(Film newFilm);
+
+    void validateFilm(Film film);
+}
